requirements.txt
keyboard
numpy
openai
openai_whisper
Pillow
pvporcupine
pvrecorder
PyAudio
pydub
pywin32
sympy
win11toast